% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AbundanceGoM}
\alias{AbundanceGoM}
\title{GoM model fit for abundance data}
\format{A list of GoM model output}
\usage{
AbundanceGoM
}
\value{
A list of GoM model output
}
\description{
GoM model fit for abundance data
}
\keyword{datasets}
