% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_omega.R
\name{compare_omega}
\alias{compare_omega}
\title{Re-ordering topic proportion matrices and Information calculation}
\usage{
compare_omega(omega1, omega2)
}
\arguments{
\item{omega1}{topic proportion fitted matrix (N x K) from model 1:}

\item{omega2}{topic proportion fitted matrix (N x K) from model 2:}
}
\description{
This function computes a re-ordering of the clusters from GoM model fit in
in one model to make it comparable with that from another. The two models are applie on the
same set of samples with same number of clusters, but the features may change from one model to another.
The two models may not be of same type as well. One could be a DAPC model, the other a standard topic model.
The idea is to check for consistency in topic proportion patterns across multiple GoM methods or across
different types of feature sets.
}
\author{
Kushal K Dey, Matthew Stephens
}

