% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countsclust.R
\name{countsclust}
\alias{countsclust}
\title{The CountsClust function- a function for clustering and visualization of counts!}
\usage{
countsclust(data, nclus_vec, samp_metadata, samp_lab_tsne, tol = 0.001,
  batch_lab, use_parallel = FALSE, use_tsne = TRUE, use_pca = TRUE,
  thresh_prop = 0, filter_prop = 0.9, expand_factor = 100,
  feature_extr_method = c("poisson", "bernoulli"), top_features = 10,
  partition = rep("TRUE", ncol(samp_metadata)), control = list())
}
\arguments{
\item{data}{counts data, with samples along the rows and features along the columns.}

\item{nclus_vec}{a vector of topics/clusters that user wants to fit.}

\item{samp_metadata}{the sample metadata, samples along the rows and each column representing some metadata information}

\item{tol}{the tolerance value for topic model fit (set to 0.001 as default)}

\item{batch_lab}{the batch labels, the output will have one Structure plot arranged by batch labels too.}

\item{use_parallel}{equal to TRUE if parallel implementation / default is FALSE}

\item{use_tsne}{equal to TRUE if the user wants to compute t-SNE / default is TRUE}

\item{use_pca}{equal to TRUE is the user wants to compute the PCA/ default is TRUE}

\item{thresh_prop}{The threshold proportion used to remove/substitute the columns of data containing NA. Deafult is 0 implying that all columns
with NA are removed.}

\item{filter_prop}{The threshold proportion that determines if the feature is too sparse or not. Default is 0.9.}

\item{expand_factor}{An expansion factor used to obtain batch corrected counts}

\item{top_features}{The number of top features per cluster that drives away that cluster from others. Default value is 10}

\item{partition}{A logical vector of same length as metadata. partition[i]=TRUE will imply that for the Structure
plot for i th metadata, no vertical line parititon between classes is used.}

\item{method}{The underlying model assumed for KL divergence measurement. Two choices considered- "bernoulli" and "poisson"}

\item{control()}{A list of control parameters for the Structure plot. The control list has the arguments
       struct.width, struct.height, cex.axis, cex.main, las.struct and las.labels.}
}
\description{
This function takes the counts data (no. of samples x no. of features), the vector of topics/clusters the user wants to fit, along
with the sample metadata and batch label information and it produces the Structure plots with and without controlling fro batch effects for
different number of topics, with arrangement determined by the sample metadata. Additionally, it outputs the cluster data output and model Bayes
factor, and also plots t-SNE and the PCA for both with and without batch cluster outputs.
}
\author{
Kushal K Dey, Matthew Stephens
}
\references{
Matt Taddy.On Estimation and Selection for Topic Models. AISTATS 2012, JMLR W\&CP 22.

            Pritchard, Jonathan K., Matthew Stephens, and Peter Donnelly. Inference of population structure using multilocus genotype data.
            Genetics 155.2 (2000): 945-959.

            Broman KW. R/qtlcharts: interactive graphics for quantitative trait locus mapping. 2015. Genetics 199:359-361

            L.J.P. van der Maaten. Accelerating t-SNE using Tree-Based Algorithms. Journal of Machine Learning Research, 2014:3221-3245
}
\keyword{Structure}
\keyword{clustering,}
\keyword{counts}
\keyword{data,}
\keyword{model,}
\keyword{plot}
\keyword{topic}

