% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.counts}
\alias{ex.counts}
\title{counts data for GTEx V6 Brain data for 200 genes}
\format{A data frame 1259 by 200 in dimensions}
\usage{
ex.counts
}
\value{
A data frame 1259 by 200 in dimensions
}
\description{
counts data for GTEx V6 Brain data for 200 genes
}
\keyword{datasets}

