% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractHighCorFeatures.R
\name{ExtractHighCorFeatures}
\alias{ExtractHighCorFeatures}
\title{Extracting most highly correlated genes with GoM topics/clusters}
\usage{
ExtractHighCorFeatures(omega, data, num_genes = 100)
}
\arguments{
\item{omega}{\eqn{\boldsymbol{omega}} matrix, the relative grades of memberships
from the GoM model fitting (a \eqn{N x K} matrix where \eqn{N} is
number of samples, \eqn{K} number of topics).}

\item{data}{\eqn{G x N} matrix of the expression profile of genes across samples,
where \eqn{G} is the number of features and \eqn{N} number of samples}

\item{num_genes}{The number of top associated genes with each cluster. Defaults to 100}
}
\value{
A list containing two items - a \eqn{K x num_genes} matrix of the
         top strongly associated/correlated indices/features for K clusters,
         and another \eqn{K x num_genes} matrix of the absolute values of the
         correlations.
}
\description{
This function compares grades of membership profile for each
 cluster in GoM model fit with the data expression profile to identify
 genes that are mostly strongly associated with each topic.
}
\examples{
data("MouseDeng2014.FitGoM")
omega_mat <- MouseDeng2014.FitGoM$clust_6$omega;
read.data1 = function() {
    x = tempfile()
   download.file('https://cdn.rawgit.com/kkdey/singleCellRNASeqMouseDeng2014/master/data/Deng2014MouseEsc.rda', destfile=x, quiet=TRUE)
   z = get(load((x)))
   return(z)
}
Deng2014MouseESC <- read.data1()
deng.counts <- Biobase::exprs(Deng2014MouseESC)
out <- ExtractHighCorFeatures(omega_mat, deng.counts, num_genes=10)
}

