% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/handleNA.R
\name{handleNA}
\alias{handleNA}
\title{Deal with NAs in the dataset!}
\usage{
handleNA(data, thresh_prop = 0)
}
\arguments{
\item{data}{counts data}

\item{thresh_prop}{threshold proportion of NAs for removal of feature or replacing the NA values}
}
\description{
This function deals with NA values in the counts dataset. If for a feature, the proportion of
NAs is greater than threshold proportion, then we remove the feature, otherwise we use MAR substitution scheme
using the distribution of the non NA values for the feature. If threshold proportion is 0, it implies removal of
all features with NA values. The default value of threshold proportion is 0.
}
\details{
This function removes NAs from the counts data
}
\examples{
mat <- rbind(c(2,4,NA),c(4,7,8),c(3,NA,NA));
handleNA(mat,thresh_prop=0.5)
handleNA(mat)
}
\author{
Kushal K Dey
}
\keyword{clustering}
\keyword{counts}
\keyword{data,}

