% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logLik.R
\name{model_logLik}
\alias{model_logLik}
\title{log likelihood of the GoM model fits}
\usage{
model_logLik(data, model_output)
}
\arguments{
\item{model_output}{\code{FitGoM} output (a \code{list}).}
}
\value{
logLik_vec a vector of GoM model log likelihoods.
}
\description{
This function takes the \code{FitGoM} output model fits
             and compute log likelihood for the GoM models.
}
\examples{

# We need both the observed count matrix and
# the GoM model fit output

# Import Deng et al data
read.data <- function() {
  x <- tempfile()
  download.file(paste0("https://cdn.rawgit.com/kkdey/",
                         "singleCellRNASeqMouseDeng2014",
                         "/master/data/Deng2014MouseEsc.rda"),
                destfile = x, quiet = TRUE)
  z <- get(load((x)))
  return(z)
  }
Deng2014MouseESC <-read.data()

# Extract observed counts
deng.counts <- Biobase::exprs(Deng2014MouseESC)

# Import GoM fitting results
data("MouseDeng2014.FitGoM")
names(MouseDeng2014.FitGoM)

model_logLik(data = deng.counts,
             model_output = MouseDeng2014.FitGoM)
}
\keyword{GoM,}
\keyword{fit}
\keyword{model}

