% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MouseDeng2014.FitGoM}
\alias{MouseDeng2014.FitGoM}
\title{GoM model fit for Deng et al 2014 single cell RNA-seq data
  on mouse}
\format{
A list of GoM model output for 6 clusters (k=2:7)
}
\usage{
MouseDeng2014.FitGoM
}
\value{
A list of GoM model output for 6 clusters (k=2:7)
}
\description{
GoM model fit for Deng et al 2014 single cell RNA-seq data
  on mouse
}
\keyword{datasets}
