% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MouseJaitinSpleen.FitGoM}
\alias{MouseJaitinSpleen.FitGoM}
\title{GoM model fit for Jaitin et al 2014 single cell RNA-seq data on mouse}
\format{A list of GoM model output for k=7}
\usage{
MouseJaitinSpleen.FitGoM
}
\value{
A list of GoM model output for k=7
}
\description{
GoM model fit for Jaitin et al 2014 single cell RNA-seq data on mouse
}
\keyword{datasets}
