% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel_GoM.R
\name{nullmodel_GoM}
\alias{nullmodel_GoM}
\title{Null models for Grade of Membership (GoM) cluster validation}
\usage{
nullmodel_GoM(counts, K, tol = 0.1, null.model = c("frequency", "richness",
  "independentswap", "trialswap"), iter_fill = 1000, iter_randomized = 100,
  plot = TRUE)
}
\arguments{
\item{counts}{The counts matrix (N x G): N- the number of samples, G- number
of features}

\item{K}{The number of clusters to fit}

\item{tol}{The tolerance of the GoM model fitted}

\item{null.model}{The type of nullmodel used (similar to the
randomizeMatrix() function argument in picante package)}

\item{iter_fill}{The number of swaps/fills in each randomized matrix build}

\item{iter_randomized}{The number of randomization matrices generated}

\item{plot}{If TRUE, plots density of log Bayes factor}
}
\value{
Returns a list with
       \item{GoMBF.obs}{log BF for the observed counts with K=2 against the
       null with no clusters}
       \item{GoMBF.rand}{a vector of log BF for each randomized count matrix
       with K=2 against the null with no clusters}
       \item{pval}{the p-value of the observed log Bayes factor against the
       ones from randomized matrices}
}
\description{
Use null models (popular in ecology) to generate randomized
matrix of counts given the observed data matrix, fit the GoM model to these
null matrices and compare the fit on null model data with that on the
observed data. Used for validating the GoM clusters
}
\examples{

data("ex.counts")
nullmodel_GoM(ex.counts,
              K=2,
              tol=500,
              null.model="frequency",
              iter_randomized=3,
              plot=FALSE)
}
