% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarHistogramStructure.R
\name{polarHistogramStructure}
\alias{polarHistogramStructure}
\title{Polar histogram designed for STRUCTURE plots}
\usage{
polarHistogramStructure(df, family = NULL, columnNames = NULL,
  binSize = 1, spaceItem = 0.2, spaceFamily = 1.2,
  familyLabelDistance = 1.2, innerRadius = 0.3, outerRadius = 1,
  guides = c(10, 20, 40, 80), alphaStart = -0.3, circleProportion = 0.8,
  direction = "inwards", familyLabels = FALSE, normalised = TRUE, palette)
}
\arguments{
\item{df}{A data.frame including columns of (family, item, score, value).
family denotes phenotype, such as tissue type. item denotes
sample lables, such as brain tissue no. 1. Score denotes cluster
memebership of each sample. value denotes cluster membership
probabilities of each sample. famiy, item, and score vectors
are required to be factor.}

\item{outerRadius}{Length of the radius to the outer edge of the circle.}

\item{guides}{A vector of values used for visual guides. Typically ranges
between 0 to 100.}

\item{circleProportion}{Percent of the ciricle (in radius) to be used in
plotting.}

\item{palette}{A vector of colors for the clusters.}

\item{Family}{A factor vector of phenotype for each sample. The level of the famliy vector determins
the order of the samples in the}
}
\value{
Returns a polar histogram.
}
\description{
This function is adapted from Christophe Ladroue's work in library(phenotypicForest)
(https://github.com/chrislad/phenotypicForest/blob/master/R/polarHistogram.R)
}
\examples{
polarHistogramStructure()
}

