% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsne_struct.R
\name{reverse_transform}
\alias{reverse_transform}
\title{Apply t-SNE on the topic proportions from structure}
\usage{
reverse_transform(x)
}
\arguments{
\item{omega}{the topic proportion matrix obtained from StructureObj/ topics ()  in maptpx}
}
\description{
This function transforms the simplex matrix omega to unconstrained matrix and then applies t-SNE (Van der Maaten) and plots the results on 2D
}
\references{
L.J.P. van der Maaten. Accelerating t-SNE using Tree-Based Algorithms. Journal of Machine Learning Research, 2014:3221-3245

           L.J.P. van der Maaten and G.E. Hinton. Visualizing High-Dimensional Data Using t-SNE. Journal of Machine Learning Research, 2008: 2579-2605.

           Broman KW. R/qtlcharts: interactive graphics for quantitative trait locus mapping. 2015. Genetics 199:359-361
}

