% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure_omega.R
\name{Structure_omega}
\alias{Structure_omega}
\title{Structure Plot of GoM model proportions (Base plot version)!}
\usage{
Structure_omega(omega, samp_metadata, batch_lab, path_struct = NULL,
  partition = rep("TRUE", ncol(samp_metadata)), control = list())
}
\arguments{
\item{omega}{the cluster membership probability/omega matrix returned by
rows and topics along columns.}

\item{samp_metadata}{the sample metadata, samples along the rows and
each column representing some metadata information
that will be used to arrange the Structure plot columns
(one plot for one arrangement).}

\item{batch_lab}{the batch labels, the output will have one Structure plot
arranged by batch labels too.}

\item{partition}{A logical vector of same length as metadata.
partition[i]=TRUE will imply that for the Structure
plot for i th metadata, no vertical line parititon
between classes is used.}

\item{path}{The directory path where we want to save the data and
Structure plots.}

\item{control()}{A list of control parameters for the Structure plot.
The control list has the arguments
struct.width, struct.height, cex.axis, cex.main,
lwd, las and color and margin parameters.}
}
\value{
Returns a non-ggplot Structure plot visualization of GoM model.
}
\description{
This function takes the cluster membership probability or omega
matrix from  topics() in maptpx or FitGoM() function in CountClust along
with sample metadata and outputs and saves a non-ggplot2 version of the
Structure plot.
}
\examples{
# Structure_omega()

}
\keyword{internal}

