% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureObj.R
\name{StructureObj}
\alias{StructureObj}
\title{Topic model fitting and Structure Plot!}
\usage{
StructureObj(data, nclus_vec, samp_metadata, tol, batch_lab, path_rds = NULL,
  path_struct = NULL, partition = rep("TRUE", ncol(samp_metadata)),
  plot = FALSE, control = list())
}
\arguments{
\item{data}{counts data, with samples along the rows and features along the columns.}

\item{nclus_vec}{the vector of clusters or topics to be fitted.}

\item{tol}{the tolerance value for topic model fit (set to 0.001 as default)}

<<<<<<< HEAD
\item{batch_lab}{the batch labels, the output will have one Structure plot arranged by batch labels too.}

\item{path_rds}{directory to store Rds object}

\item{path_struct}{path to store structure plot}

\item{partition}{A logical vector of same length as metadata. partition[i]=TRUE will imply that for the Structure
plot for i th metadata, no vertical line parititon between classes is used.}

\item{plot}{TRUE or FALSE. To make the Structure plots or not.}

\item{control()}{A list of control parameters for the Structure plot. The control list has the arguments
struct.width, struct.height, cex.axis, cex.main, las.struct, lwd, las and color and margin parameters.}
=======
\item{path}{The directory path where we want to save the data and Structure plots.}
>>>>>>> upstream/master
}
\description{
This function takes the counts data (no. of samples x no. of features) and the value of K, the number of topics or
cluster to fit, along with sample metadata information and fits the topic model (due to Matt Taddy, check package
maptpx) and outputs the Structure plot with the column bars in
Structure plot arranged as per the sample metadata information.
}
\author{
Kushal K Dey, Matt Taddy, Matthew Stephens, Ida Moltke
}
\references{
Matt Taddy.On Estimation and Selection for Topic Models. AISTATS 2012, JMLR W\&CP 22.

           Pritchard, Jonathan K., Matthew Stephens, and Peter Donnelly. Inference of population structure using multilocus genotype data.
           Genetics 155.2 (2000): 945-959.
}
\keyword{Structure}
\keyword{clustering,}
\keyword{counts}
\keyword{data,}
\keyword{plot}

