% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsne_struct.R
\name{tsne_struct}
\alias{tsne_struct}
\title{Apply t-SNE on the topic proportions from structure}
\usage{
tsne_struct(omega, samp_metadata, path)
}
\arguments{
\item{omega}{the topic proportion matrix obtained from StructureObj/ topics ()  in maptpx}

\item{samp_metadata}{The sample labels used for identifying labels in tSNE plot}

\item{path}{The path where you want to save the t-SNE plot}
}
\description{
This function transforms the simplex matrix omega to unconstrained matrix and then applies t-SNE (Van der Maaten) and plots the results on 2D
}
\author{
Kushal K Dey
}
\references{
L.J.P. van der Maaten. Accelerating t-SNE using Tree-Based Algorithms. Journal of Machine Learning Research, 2014:3221-3245

           L.J.P. van der Maaten and G.E. Hinton. Visualizing High-Dimensional Data Using t-SNE. Journal of Machine Learning Research, 2008: 2579-2605.

           Broman KW. R/qtlcharts: interactive graphics for quantitative trait locus mapping. 2015. Genetics 199:359-361
}
\keyword{model}
\keyword{topic}
\keyword{tsne,}

