% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BatchCorrectedCounts.R
\name{BatchCorrectedCounts}
\alias{BatchCorrectedCounts}
\title{Obtaining Batch effect Corrected counts!}
\usage{
BatchCorrectedCounts(data, expand_factor, batch_lab)
}
\arguments{
\item{data}{counts data, with samples along the rows and features along the columns.}

\item{expand_factor}{a expanding factor required to convert batch corrected CPM values to batch corrected counts (default=100)}

\item{batch_lab}{the batch label vector}
}
\description{
This function first converts counts data to CPM data, then apply a linear model with the batch effect as a factor. The batch corrected
             CPM residuals are then exponentiated and multiplied by expand_factor. These values for each sample and each feature is taken as the
             mean for a Poisson random number generation. The obtained counts are then devoid of batch effects.
}
\author{
Kushal K Dey, Joyce HSiao
}
\keyword{batch}
\keyword{counts}
\keyword{data,}
\keyword{effect}

