% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compGoM.R
\name{compGoM}
\alias{compGoM}
\title{compGoM: compare GoM model fits across K or across different runs
through log-likelihood, BIC and null loglikelihood}
\usage{
compGoM(data, model)
}
\arguments{
\item{data}{matrix on which GoM model is fitted (samples along rows, genes
along columns)}

\item{model}{\code{FitGoM} or\code{maptpx::topics} function output
(either a class \code{topics} or a  \code{list} of class
 \code{topics}).}
}
\value{
compGoM_models a vector list that returns the  BIC and loglikelihood
               values for each of the fitted models in \code{model}.
}
\description{
This function takes the \code{FitGoM}/maptpx fitted model
             and computes log likelihood, BIC and null model loglikelihood
             for the fitted GoM models.
}
\examples{

read.data <- function() {
  x <- tempfile()
  download.file(paste0("https://cdn.rawgit.com/kkdey/",
                         "singleCellRNASeqMouseDeng2014",
                         "/master/data/Deng2014MouseEsc.rda"),
                destfile = x, quiet = TRUE)
  z <- get(load((x)))
  return(z)
  }
Deng2014MouseESC <-read.data()

# Extract observed counts
deng.counts <- Biobase::exprs(Deng2014MouseESC)

# Import GoM fitting results
data("MouseDeng2014.FitGoM")
names(MouseDeng2014.FitGoM)

compGoM(data = t(deng.counts),
           model = MouseDeng2014.FitGoM)
compGoM(data = t(deng.counts),
           model = MouseDeng2014.FitGoM$clust_3)

}
\keyword{GoM,}
\keyword{fit}
\keyword{model}
