% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compGoM.R
\name{compGoM}
\alias{compGoM}
\title{Compare GoM model fits across K or across different runs
  through log-likelihood, BIC and null loglikelihood}
\usage{
compGoM(data, model)
}
\arguments{
\item{data}{Matrix on which GoM model is fitted (samples along
rows, genes along columns)}

\item{model}{\code{FitGoM} or\code{maptpx::topics} function output
(either a class \code{topics} or a \code{list} of class
\code{topics}).}
}
\value{
compGoM_models a vector list that returns the BIC and
  loglikelihood values for each of the fitted models in \code{model}.
}
\description{
This function takes the \code{FitGoM}/maptpx fitted
  model and computes log likelihood, BIC and null model loglikelihood
  for the fitted GoM models.
}
\examples{
# Extract observed counts
library(singleCellRNASeqMouseDeng2014)
data("Deng2014MouseEsc")
deng.counts <- exprs(Deng2014MouseESC)

# Import GoM fitting results
data("MouseDeng2014.FitGoM")
names(MouseDeng2014.FitGoM)

compGoM(data = t(deng.counts),model = MouseDeng2014.FitGoM)
compGoM(data = t(deng.counts),model = MouseDeng2014.FitGoM$clust_3)

}
\keyword{GoM,}
\keyword{fit}
\keyword{model}
