% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTopFeatures.R
\name{ExtractTopFeatures}
\alias{ExtractTopFeatures}
\title{extractTopFeatures- extracting top driving genes driving GoM clusters}
\usage{
ExtractTopFeatures(theta, top_features = 10, method = c("poisson",
  "bernoulli"), options = c("min", "max"))
}
\arguments{
\item{theta}{The theta matrix obtained from the topic model fitting (it is a G x K matrix where G is number of features, K number of topics)}

\item{top_features}{The number of top features per cluster that drives away that cluster from others. Default value is 10}

\item{method}{The underlying model assumed for KL divergence measurement. Two choices considered- "bernoulli" and "poisson"}

\item{options}{if "min", for each cluster k, we select features that maximize the minimum KL divergence of
cluster k against all other clusters for each feature. If "max", we select features  that maximize the
maximum KL divergence of cluster k against all other clusters for each feature.}
}
\value{
A matrix (K x top_features) which tabulates in k th  row the top feature indices driving the cluster k.
}
\description{
This function uses relative expression profile of the GoM clusters for each feature, and applies a KL divergence mechanism to obtain
a list of top features that drive the clusters.
}
\examples{
theta_mat <- MouseDeng2014.topicFit$clust_6$theta;
top_features <- ExtractTopFeatures(theta_mat, top_features=100,
                                  method="poisson", options="min");
gene_list <- do.call(rbind, lapply(1:dim(top_features)[1],
                                  function(x) deng.gene_names[top_features[x,]]))

}
\author{
Kushal K Dey, Matthew Stephens
}

