% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTopFeatures.R
\name{ExtractTopFeatures}
\alias{ExtractTopFeatures}
\title{The CountsClust function- a function for clustering and visualization of counts!}
\usage{
ExtractTopFeatures(theta, top_features = 10, method = c("poisson",
  "bernoulli"))
}
\arguments{
\item{theta}{The theta matrix obtained from the topic model fitting (it is a G x K matrix where G is number of features, K number of topics)}

\item{top_features}{The number of top features per cluster that drives away that cluster from others. Default value is 10}

\item{method}{The underlying model assumed for KL divergence measurement. Two choices considered- "bernoulli" and "poisson"}
}
\description{
This function uses the relativ expression profile of the clusters across the features, and applies a KL divergence mechanism to obtain
a list of features that are significant in driving the clusters. Often annotating these features may give us an idea as to whether there is a common
property among the features for each cluster.
}
\author{
Kushal K Dey, Matthew Stephens
}

