% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGoM.R
\name{FitGoM}
\alias{FitGoM}
\title{Grade of Membership (GoM) model fit !}
\usage{
FitGoM(data, K, tol = 0.1, path_rda = NULL, control = list())
}
\arguments{
\item{data}{counts data, with samples along the rows and features
along the columns.}

\item{K}{the vector of clusters or topics to be fitted.}

\item{tol}{Tolerance value for GoM model log posterior increase
at successive iterations (set to 0.1 as default).}

\item{path_rda}{The directory path for saving the GoM model output.
If NULL, it will return the output to console.}

\item{control}{Control parameters. Same as topics() function of
maptpx package.}
}
\value{
Saves the GoM model fit output for each cluster in vector K at the
               directory path in path_rda.
}
\description{
Fits a grade of membership model to count data. Default input
               includes a sample-by-feature matrix, the number of clusters
               (topics) to fit (K). The function is a wrapper of the topics()
               function implemented in Matt Taddy's maptpx pacakge.
}
\examples{

data("ex.counts")
out <- FitGoM(ex.counts, K=4, tol=100, control=list(tmax=100))

}
\references{
Matt Taddy. On Estimation and Selection for Topic Models.
               AISTATS 2012, JMLR W\&CP 22.

           Pritchard, Jonathan K., Matthew Stephens, and Peter Donnelly.
               Inference of population structure using multilocus genotype
               data. Genetics 155.2 (2000): 945-959.
}
\keyword{Structure}
\keyword{clustering,}
\keyword{counts}
\keyword{data,}
\keyword{plot}

